# openapi.yaml
openapi: 3.0.3
info:
  title: Área de Membros — Thiago Cantalovo
  version: "1.0.0"
  description: >
    API REST para integrações externas (n8n, Make, Zapier, Flowise, Langflow, checkouts, etc.).
    Documentação em prosa equivalente: API.md. Mantida manualmente — não é gerada a partir do código.
servers:
  - url: https://membros.thiagocantalovo.com
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/admin/usuarios/{id}/resetar-senha:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [Membros]
      summary: Solicitar redefinição de senha para um membro existente
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/produtos:
    get:
      tags: [Produtos]
      summary: Listar produtos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Produtos]
      summary: Criar produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content_type]
              properties:
                title: { type: string }
                description: { type: string }
                content_type: { type: string, enum: [video, pdf, curso] }
                content_url: { type: string, nullable: true }
                banner_url: { type: string, nullable: true }
                is_pack: { type: boolean, default: false }
                sort_order: { type: integer, default: 0 }
                is_active: { type: boolean, default: true }
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/produtos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Produtos]
      summary: Buscar produto
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Produtos]
      summary: Editar produto
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                content_type: { type: string, enum: [video, pdf, curso] }
                content_url: { type: string, nullable: true }
                banner_url: { type: string, nullable: true }
                is_pack: { type: boolean }
                is_active: { type: boolean }
                sort_order: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Produtos]
      summary: Excluir produto
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/admin/certificados:
    get:
      tags: [Certificados]
      summary: Listar certificados emitidos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificates:
                    type: array
                    items: { $ref: "#/components/schemas/Certificate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Certificados]
      summary: Emitir certificado manualmente (ignora a exigência de conclusão do curso)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccessRef" }
      responses:
        "200":
          description: Já existia um certificado para este par membro/produto — retorna o existente
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificate: { $ref: "#/components/schemas/Certificate" }
        "201":
          description: Certificado novo emitido
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificate: { $ref: "#/components/schemas/Certificate" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/webhook/asaas:
    post:
      tags: [Webhooks de Entrada]
      summary: Notificação de pagamento do Asaas
      security: []
      parameters:
        - name: asaas-access-token
          in: header
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string, example: PAYMENT_CONFIRMED }
                payment:
                  type: object
                  properties:
                    id: { type: string }
                    customer: { type: string }
                    value: { type: number }
                    externalReference:
                      type: string
                      description: UUID do produto na plataforma
      responses:
        "200":
          description: Sempre 200 — falhas de processamento ficam em webhook_logs, não na resposta HTTP
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: x-api-key ausente ou inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    BadRequest:
      description: Dados inválidos ou faltando na requisição
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        content_type: { type: string, enum: [video, pdf, curso] }
        content_url: { type: string, nullable: true }
        banner_url: { type: string, nullable: true }
        is_pack: { type: boolean }
        is_active: { type: boolean }
        is_featured: { type: boolean }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
    AccessRef:
      type: object
      required: [user_id, product_id]
      properties:
        user_id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
...
    Certificate:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        issued_at: { type: string, format: date-time }